liquidation-bot/requirements.txt
web3>=6.11.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
